use more int64_t

+++ sperl_heap.c
@@ -4,9 +4,9 @@
 #include "sperl_heap.h"
 #include "sperl_allocator_util.h"
 
-void* SPerl_HEAP_alloc(SPerl* sperl, int32_t size) {
+void* SPerl_HEAP_alloc(SPerl* sperl, int64_t size) {
   (void) sperl;
-  if (size < 0 || SIZE_MAX < (uint32_t) size) {
+  if (size < 0 || SIZE_MAX < (int64_t) size) {
     SPerl_ALLOCATOR_UTIL_exit_with_malloc_failure();
   }
   return SPerl_ALLOCATOR_UTIL_safe_malloc(size, sizeof(uint8_t));
